+++ src/Akeneo/Pim/Structure/Bundle/Query/PublicApi/Attribute/Sql/SqlGetUniqueAttributeCodes.php
@@ -17,6 +17,9 @@ final class SqlGetUniqueAttributeCodes implements GetUniqueAttributeCodes
         $this->connection = $connection;
     }
 
+    /**
+     * @return string[]
+     */
     public function all(): array
     {
         $query = <<<SQL